requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.0
contourpy==1.2.1
cycler==0.12.1
fastapi==0.115.12
fonttools==4.51.0
h11==0.16.0
idna==3.10
kafka-python==2.2.6
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.26.4
opencv-contrib-python==4.9.0.80
packaging==24.0
pandas==2.2.3
pillow==10.3.0
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2025.1
setuptools==69.2.0
six==1.16.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.1
uvicorn==0.34.2
wheel==0.43.0
